requirements.txt
fastapi==0.118.0
uvicorn==0.41.0
sqlalchemy==2.0.44
alembic==1.17.1
PyMySQL==1.1.2
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-dotenv==1.1.1
email-validator==2.3.0
